+++ pkg/providers/appsec/testdata/TestDSURLProtection/url_protection_rule_id_missing_match_by_id.tf
@@ -1,0 +1,9 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+data "akamai_appsec_url_protection_rule" "test" {
+  config_id = 43007
+}
+